fix sendrecv() without interim poll
There was a "clever" optimization that skipped sendrecv() handling
when the transfer's pollset was empty. This happens for paused
transfers, for example.

Unforunately, if the libcurl application never calls curl_multi_poll(),
the pollset is and will aways remain empty, prevent the transfer from
progressing.

Remove this "optimization" and always try send/receive where applicable.