cmake: move examples to a separate target
Also make `CURL_EXAMPLES` default to `ON`. It means to generate the
rules for examples by default, but to build them. An explicit
`make curl-examples` (or ninja, etc) command is necessary to build them.

This syncs behaviour with autotools, and also how both are building
tests.

It allows to move examples in separate CI steps to measure them and move
them around to not slow down iteration. (After pending build speed-ups,
building examples became the longest running part of a build.)