DEPRECATE: remove hyper in January 2025
Previously this document stated we would do it after February, but now
it will be done already for the first January 2025 release.

The reason being that since we decated to deprecate hyper, the
degradation speed as only increased as now no one bothers to fix issues
in the hyper side of things. Also: not a single soul has yet spoken up
in favor of keeping the support.