CI: log platform info in pytest
Use the pytest-astropy-header package to log information about the
running platform in pytest invocations so Test Clutch can include it in
the test run metadata.  This also logs some things less relevant to
curl, but it's small and simple and provides the information we want and
since it's only running in the CI build it won't generally be seen by
humans.

Closes #15256