smb: do not redefine `getpid` on Windows
Make an `#if` branch instead for the single place it was used.

Redefining symbols can backfire if that symbol is used in system
headers, especially in unity build. We haven't seen a fallout in
CI or supported envs, but do it anyway for good measure.

Bug report: https://datagirl.xyz/posts/wolfssl_curl_w2k.html