workaround for gnutls 1.0/1.1 handshake failure
Not sure how ubuntu manages to system-wide disable these and
we are not able to override it. When we set a gnutls priority
like:

NORMAL:-ARCFOUR-128:-CTYPE-ALL:+CTYPE-X509:-VERS-SSL3.0:-VERS-TLS-ALL:+VERS-TLS1.0

gnutls does fail at *starting* the handshake, e.g. it does not
send a ClientHello. So this has nothing to do with the server.
There just seem to be any algorithms disabled that are needed
got TLS1.0 and TLS1.1.