h2, auto reset on server eos
When a server signals EOS from its side and the curl
upload is unfinished and the server has not given a
positive HTTP status response, auto RST the stream to
signal that the upload is incomplete and that the whole
transfer can be stopped.

refs #15316, where the server responds with 413 on an
upload but does not RST the stream from its side, as
httpd and others do.