GHA/windows: increase timeout for vcpkg jobs
The openssl job in particular no longer fits into 10 minutes since
the 20241015.1.0 GHA windows-latest image update.
The `run tests` step takes ~10 minutes now, up from ~4. That is
6 minutes more than before these updates.

c-ares got bumped from 1.33.1 to 1.34.1 in vcpkg and the affected
vcpkg job is the only one using this dependency.

It may be the reason for this regression.

Before:
https://github.com/curl/curl/actions/runs/11386748199/job/31679733295
https://github.com/curl/curl/actions/runs/11347976608/job/31560690219

After:
https://github.com/curl/curl/actions/runs/11462332743/job/31893491625?pr=15364

Ref: https://github.com/curl/curl/pull/15335#issuecomment-2423759953