schannel: check for close notify on ssl shutdown
- Revert "schannel: ignore error on recv beyond close notify".

- During Schannel SSL shutdown do not error if schannel_recv failed but
  the close notify flag was set.

This is a revert of the change earlier today so that we again error on a
recv call for more bytes after a close notify has been received. The
change to ignore the error was made to allow SSL shutdown to call
schannel_recv with 1024 bytes and not have it error, however in other
cases attempting to read more data should remain an error. Therefore
this is best addressed in SSL shutdown instead.

Closes #xxxx