macos: disable `availability` workaround for gcc 14.2.0+
Homebrew gcc 14.2.0_1 fixed the issue, and the workaround is no longer
needed. Not only not needed, but the workaround breaks the fixed
version. There is no way to detect if gcc is 14.2.0 or 14.2.0_1, so
this patch drops the local workaround for both 14.2.0 releases.

It means the issue will happen with 14.2.0, in which case the solution
is to update to 14.2.0_1.

Sadly it doesn't seem possible to fix both 14.2.0 and 14.2.0_1. (Perhaps
unless doing an elaborate env detection, but that seems overkill for
this single transition.)

This issue did not yet hit CI, because the macos image is still coming
with 14.2.0, but will hit once it's bumped to 14.2.0_1.

Follow-up to e91fcbac7d86292858718a0bfebad57978761af4 #14155