GHA/macos: make gcc dictate the configured SDK
As discovered earlier, Homebrew gcc is built against a specific Apple
SDK and doesn't work when matched up with a different SDK, e.g. the
one advertised as default by the macos runner image.

Before this patch this was resolved with brute force by zapping the
hack-layer gcc component to avoid the bad interference. This worked
for us, but it's fragile, accidental and doesn't translate to any
real-world build environment. Thus, impractical.

This patch avoids this by explicitly configuring the SDK gcc was built
for and meant to be used with.

It assumes that the gcc binaries preinstalled on the runner images
always ship with the SDK version they reference. It's non-trivial to
guarantee that given the different build/release/update cycles of the
projects involved, but since this is the only officially supported
way to use them, hopefully this is ensured to be kept in sync.
Since we don't do `brew update`, there is emphasis on preinstalled.

In curl-for-win CI jobs we do `brew update` and expect that scenario
to work too.

Follow-up to c349bd668c91f2484ae21c0f361ddf497143093c #14097