cmake: do not use `getpass_r` with Apple
It's not offered by Apple operating systems, but in some cases it gets
mis-detected when using `check_function_exists()`. Then making the build
fail later.

A better fix would be to figure out why are these mis-detections. One
common pattern seen is this option, but it may or not be the cause:
```
-DCMAKE_TOOLCHAIN_FILE=[...]/toolchains/ios.toolchain.cmake
```
Ref: #15525

Regression from 8e345057761a8f796403923a96f2c8fd3edca647 #15164

Reported-by: Dan Rosser
Fixes #15557
Closes #15558