Document default `User-Agent`
curl offers a `--user-agent` option for modifying the `User-Agent`
header supplied in its requests.

The man page section for this option explains how to use the
`--user-agent` option, but does not explain which `User-Agent` curl uses
by default.

By default, curl uses curl/VERSION, such as `User-Agent: curl/8.11.0`.

Note that this appears to be different from the libcurl default
(no User-Agent header).

This commit will document the default `User-Agent` in the man page
section for the `--user-agent` option, as well as on the
"Art of Scripting" page.

The `%VERSION` placeholder will be used to insert the current version
as described in the man page generator docs.