configure: drop unused openssl feature tests HAVE_X509_H, HAVE_CRYPTO_H, HAVE_SSL_H
The logic was doing extra work when openssl/x509.h was not found.

`AC_CHECK_HEADERS()` already considered the detection successful
if any other headers were found. So the logic kicked in in case
no headers were found. In such case it considered OpenSSL be found
_internally_ if three of the headers were found without the
`openssl/` prefix, but in this case `USE_OPENSSL` was not promoted
to `curl_config.h`. Meaning the logic did not work even if it could
detect OpenSSL with the unprefixed headers.

Logic added here d99c20f62876457ca6880d706825e68f695bda18 (2008)

At the time, there was an extra `AC_DEFINE(USE_SSLEAY, 1 ...` logic
after this code, which kicked in in the fallback case, but that
code was deleted in 709cf76f6bb7dbaca14e3e8df160ccfac04dcecb (2015)