digest: change cnonce length from 32 to 16
The 'cnonce' length is not specified. Other programs (Postname, Chrome,
Python request) does use the small 'cnonce' length. In order to be
consistent with the other programs, the value is therefore also set at 16.

The reason for the change was, that silly implementation cannot handle
'cnonce' values greater than '16'. A connection via 'digest' auth could not
be established with this system, because curl is using the 'cnonce' value '32'.
So that these faulty implementation also work with 'curl' and not only with
the other mentioned programs, the 'cnonce' value must be changed to '16'.

Closes: 15653

Signed-off-by: Florian Eckert <fe@dev.tdt.de>