Fix client mime reader stall on small read lengths
The base64 mime encoder stalls when it cannot encode a full
3 byte input set into the read buffer. The workaround for this
limitation was incomplete and could lead to stalled transfers
when the last chunk to upload was smaller than 4 bytes.

Use a tmp buffer on small reads to allow mime encoders more
space to put their things.

Add test case reproducing the issue and fix.

refs #15688