using "session_key" for lookups into TLS session cache
"session_key" is a printable string catching all the relevant
information to identify peer and ssl settings of a connection
filter. It does *not* contain private information. If a connection
uses SRP or a client certificate, this is additionally matched
and not part of the session key.

Implemented use in gnutls backend. More to follow.