rename (again) and separation of data vs object
rename 'ssl_spool' to 'ssl_scache' since we talk about
the session cache already a lot.

separate calls for data and object storage (with schannel being
hte only user of that). Because we have a hash lookup now and
getting data bytes when expecting an object instance will only
lead to crashes or other nastiness.