netrc: fix password-only entries
When a specific hostname matched, and only a password is set before
another machine is specified in the netrc file, the parser would not be
happy and stop there and return the password-only state. It instead
continued and did not return a match.

Added test 2005 to verify this case

Reported-by: Ben Zanin
Fixes #15767