fixup better expire checking
The test script now generates TWO numbers for each %days[], and the
function that compares and verifies output is fine with either of the
two numbers.

This, so if the test case is generated the second before curl runs, that
updated expiry number is also deemed okay. It still checks for an exact
match of either number.