cmake: drop VS2010 workaround
Added in a94a68a3c1d04ccb53e46baa69753bbf6354ee14, without details,
referring to a CMake bug, but without a version number. The latst
version was 2.8.10 at the time. This is no longer supported with
curl. It seems safe to assume this has been fixed in the oldest
CMake version supported by curl (3.7) or in a newer version up to
v3.24 which was the last CMake supporting VS2010 natively.