cm: cegcc -> mingw32ce
The CeGCC project offers two modes, one with an extra layer of
unixy stuff (with errno for example), via cegcc.dll, and one other,
bare WinCE mode (without errno).

In this patch I focus on the bare mode. The cegcc mode is also missing
from the binary distros.

Make the naming reflect that.

https://cegcc.sourceforge.net/docs/using.html